requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
langchain>=0.2.0
langgraph>=0.2.0
pydantic>=2.5.0
httpx>=0.25.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
pytest>=7.4.0
pypdf>=3.17.0
tiktoken>=0.5.0
numpy>=1.24.0
nltk>=3.8.0
openai>=1.12.0